Move ChurchSettings to account-management section
